+++ DesignerMacros/src/services-cqrs-crud/repository-operation-api/createCqrs.ts
@@ -1,8 +1,8 @@
 /// <reference path="../../common/openSelectItemDialog.ts" />
 /// <reference path="../../common/repositoryServiceHelper.ts" />
+/// <reference path="common.ts" />
 
-async function execute(repositoryOperation: MacroApi.Context.IElementApi) {
-
+async function createCQRSService(repositoryOperation: IElementApi, diagram: IDiagramApi | undefined) {
     let servicePackages = getPackages().filter(pkg => pkg.specialization === "Services Package");
     let selectedPackage: MacroApi.Context.IPackageApi;
     if (servicePackages.length == 1) {
@@ -17,16 +17,21 @@ async function execute(repositoryOperation: MacroApi.Context.IElementApi) {
 
     RepositoryServiceHelper.createCqrsAction(repositoryOperation, folder, true);
 
-    // this can only be triggered from the diagram, so there will always be one.
-    const diagram = getCurrentDiagram();
-    
+    if (diagram == null) {
+        diagram = await getOrCreateDiagram(folder, repositoryOperation, "CQRS Creation Options");
+    }
+
+    if (diagram == null) {
+        return;
+    }
+
     //Since we're adding a single new element on the diagram, it may not be positioned below the last created one.
     let lastActionVisual: MacroApi.Context.IElementVisualApi = null;
     for (let action of folder.getChildren("Command").concat(folder.getChildren("Query"))) {
         if (diagram.isVisual(action.id)) {
             var actionElement = diagram.getVisual(action.id);
 
-            if(!lastActionVisual || actionElement.getPosition().y > lastActionVisual.getPosition().y) {
+            if (!lastActionVisual || actionElement.getPosition().y > lastActionVisual.getPosition().y) {
                 lastActionVisual = actionElement;
             }
         }
@@ -41,22 +46,14 @@ async function execute(repositoryOperation: MacroApi.Context.IElementApi) {
             x: repoElement.getPosition().x - (repoElement.getSize().width / 1.5),
             y: lastActionVisual.getPosition().y + (lastActionVisual.getSize().height * 1.5)
         };
-    }else {
-        if(diagram.isVisual(repository.id)){
+    } else {
+        if (diagram.isVisual(repository.id)) {
             newPosition = {
                 x: repoElement.getPosition().x - (repoElement.getSize().width / 1.5),
                 y: repoElement.getPosition().y
             };
         }
     }
-   
+
     diagram.layoutVisuals(folder, newPosition, true);
 }
-
-/**
- * Used by Intent.Modelers.Services.DomainInteractions
- *
- * Source code here:
- * https://github.com/IntentArchitect/Intent.Modules/blob/development/DesignerMacros/src/services-cqrs-crud/create-cqrs-repository-operation-macro-advanced-mapping/create-cqrs-repository-operation-macro-advanced-mapping.ts
- */
-//await execute(element);